some change in controller

+++ src/controllers/trustBill.controller.ts
@@ -1,5 +1,6 @@
 import { Request, Response } from "express";
 import TrustBill from "../models/trust.bill";
+import mongoose from "mongoose";
 
 export const createTrustBill = async (req: Request, res: Response) => {
     try {
@@ -39,12 +40,43 @@ export const getTrustBill = async (req: Request, res: Response) => {
     }
 }
 
+// export const trustDeleteById = async (req: Request, res: Response) => {
+//     const id = req.params.id
+//     try {
+//         await TrustBill.findByIdAndDelete(id)
+//         res.status(200).json({ success: true, message: "entry delete successful" })
+//     } catch (error) {
+//         res.status(500).json({ success: false, message: (error as Error).message })
+//     }
+// }
+
 export const trustDeleteById = async (req: Request, res: Response) => {
-    const id = req.params.id
-    try {
-        await TrustBill.findByIdAndDelete(id)
-        res.status(200).json({ success: true, message: "entry delete successful" })
-    } catch (error) {
-        res.status(500).json({ success: false, message: (error as Error).message })
+    const { id } = req.params;
+
+    if (!mongoose.Types.ObjectId.isValid(id)) {
+        res.status(400).json({ success: false, message: "Invalid ID" });
+        return
     }
-}
+
+    try {
+        const deleted = await TrustBill.findByIdAndDelete(id);
+
+        if (!deleted) {
+            res.status(404).json({
+                success: false,
+                message: "Record not found"
+            });
+            return
+        }
+
+        res.status(200).json({
+            success: true,
+            message: "Entry deleted successfully"
+        });
+    } catch (error) {
+        res.status(500).json({
+            success: false,
+            message: (error as Error).message
+        });
+    }
+};
